more precise timing
renamed Time::elapsed -> Time::seconds, change type to double
updated game to store ticks (microseconds) instead of milliseconds for better update precision

+++ include/blah/input/virtual_stick.h
@@ -70,9 +70,9 @@ namespace Blah
 		bool m_released = false;
 
 		float m_i_deadzone;
-		float m_last_press_time = -1;
-		float m_last_release_time = -1;
-		float m_repeat_press_time = -1;
+		double m_last_press_time = -1;
+		double m_last_release_time = -1;
+		double m_repeat_press_time = -1;
 
 	public:
 		VirtualStick();
